--- src/pages/edit.tsx
@@ -1,295 +1,0 @@
-import Link from 'next/link';
-import {
-  Box,
-  Container,
-  Flex,
-  Grid,
-  Heading,
-  Spacer,
-  Tab,
-  TabList,
-  TabPanel,
-  TabPanels,
-  Tabs,
-  useToast,
-} from '@chakra-ui/react';
-import { useEffect, useState } from 'react';
-import MemberForm from '../components/MemberForm';
-import { getAllMembers } from '../firebase/member';
-import { Member } from '../fixtures/Members';
-import MembersTable from '../components/MembersTable';
-import PairingForm from '../components/PairingForm';
-import PairingsTable from '../components/PairingsTable';
-import { Pairing } from '../fixtures/Pairings';
-import { getAllPairings } from '../firebase/pairings';
-import { GetStaticProps } from 'next';
-import Splash from '../components/Splash';
-import GradientButton from '../components/GradientButton';
-import { deleteMember, updateMember } from '../client/membersService';
-import useAuth from '../hooks/useAuth';
-
-// TODO: organize imports
-
-interface EditPageProps {
-  members: Member[];
-  pairings: Pairing[];
-}
-
-/**
- * Page to view and edit the PSA member database
- */
-const EditPage = ({ members, pairings }: EditPageProps) => {
-  const { loadSplash } = useAuth();
-  const [membersList, setMembersList] = useState<Member[]>([]);
-  const [pairingsList, setPairingsList] = useState<Pairing[]>([]);
-  const [loadingPairs, setLoadingPairs] = useState<boolean>(false);
-  const [navDisabled, setNavDisabled] = useState<boolean>(false);
-  const toast = useToast();
-
-  // Load members and pairings lists from props upon render
-  useEffect(() => {
-    setMembersList(members);
-    setPairingsList(pairings);
-  }, []);
-
-  /**
-   * Refresh the members list from database
-   */
-  const refreshMembers = async () => {
-
-    const fetchedMembers = await (await fetch(`/api/members`)).json();
-
-    setMembersList(fetchedMembers);
-    return fetchedMembers;
-  };
-
-  /**
-   * Refresh the pairings list from database
-   */
-  const refreshPairings = async () => {
-    setLoadingPairs(true);
-
-    const fetchedPairings = await (await fetch(`/api/pairings`)).json();
-
-    setPairingsList(fetchedPairings);
-    setLoadingPairs(false);
-    return fetchedPairings;
-  };
-
-  /**
-   * Refresh both members and pairings tables
-   */
-  const refreshTables = async () => {
-    await refreshMembers();
-    await refreshPairings();
-  };
-
-  /**
-   * Callback to update member's data in state and database
-   * @param updated member with new data
-   * @param i index of member in state
-   */
-  const changeMember = async (updated: Member, i: number) => {
-    // If no change is present, skip
-    const original = membersList[i];
-    if (updated.name === original.name && updated.classOf === original.classOf)
-      return;
-
-    // Check schema
-    if (!Boolean(updated.name) || updated.classOf.length !== 4) {
-      toast({
-        title: 'Error',
-        status: 'error',
-        description:
-          'Name must be non-empty. Class must be exactly 4 digits long. Please refresh.',
-      });
-      return;
-    }
-
-    // Update member in database
-    const { success, message, member } = await updateMember(updated);
-
-    // Send toast
-    toast({
-      title: success ? 'Success!' : 'Error',
-      status: success ? 'success' : 'error',
-      description: message,
-    });
-
-    if (success) {
-      // Update member in state
-      const updatedMembers = [...membersList];
-      updatedMembers[i] = member!;
-      setMembersList(updatedMembers);
-
-      // Update affected pairings
-      const updatedPairings = [...pairingsList];
-      updatedPairings.forEach((p) => {
-        if (p.ak.id === updated.id) {
-          p.ak = member!;
-        } else if (p.ading.id === updated.id) {
-          p.ading = member!;
-        }
-      });
-      setPairingsList(updatedPairings);
-    }
-  };
-
-  /**
-   * Callback to delete member from database and state
-   * @param member member to delete
-   */
-  const removeMember = async (member: Member) => {
-    if (!member.id) return;
-
-    // Delete member from database
-    const response = await deleteMember(member.id);
-
-    if (response.success) {
-      toast({
-        title: 'Deletion complete',
-        status: 'info',
-        description: `${member.name} has been deleted.`,
-      });
-    } else {
-      toast({
-        title: 'Error',
-        status: 'error',
-        description: response.message,
-      });
-    }
-
-    // Get member documents to update
-    const memberIdsToDecrementAks: string[] = [];
-    const memberIdsToDecrementAdings: string[] = [];
-    const filteredPairings = pairingsList.filter((p) => {
-      if (p.ak.id === member.id) {
-        memberIdsToDecrementAks.push(p.ading.id);
-        return false;
-      } else if (p.ading.id === member.id) {
-        memberIdsToDecrementAdings.push(p.ak.id);
-        return false;
-      }
-      return true;
-    });
-
-    const convertMemberIdToIndex = (id: string) =>
-      membersList.findIndex((m) => m.id === id);
-    const memberIndexesToDecrementAks = memberIdsToDecrementAks.map(
-      convertMemberIdToIndex
-    );
-    const memberIndexesToDecrementAdings = memberIdsToDecrementAdings.map(
-      convertMemberIdToIndex
-    );
-
-    // Update member documents
-    const updatedMembers = [...membersList];
-    for (const index of memberIndexesToDecrementAks) {
-      updatedMembers[index].aks -= 1;
-    }
-    for (const index of memberIndexesToDecrementAdings) {
-      updatedMembers[index].adings -= 1;
-    }
-
-    // Set state to updated documents
-    setMembersList(updatedMembers.filter(({ id }) => id !== member.id));
-    setPairingsList(filteredPairings);
-  };
-
-  if (loadSplash) {
-    return <Splash />;
-  }
-
-  return (
-    <Box>
-      <Flex p={4} align="center" borderColor="white">
-        <Heading as="h1" fontSize="2xl" fontWeight="light">
-          <Link href="/">Lineage</Link>
-        </Heading>
-        <Spacer />
-        <GradientButton
-          disabled={navDisabled}
-          onClick={() => setNavDisabled(true)}
-        >
-          <Link href="/lineages">
-            {navDisabled ? 'Loading...' : 'View lineages'}
-          </Link>
-        </GradientButton>
-      </Flex>
-      <Grid templateColumns={['100%', '50% 50%']}>
-        {/* Member column */}
-        <Container>
-          <Heading variant="h3" my={5}>
-            Add member
-          </Heading>
-          <MemberForm members={membersList} setMembers={setMembersList} />
-        </Container>
-
-        {/* Pairing column */}
-        <Container>
-          <Heading variant="h3" my={5}>
-            Add pairing
-          </Heading>
-          <PairingForm
-            members={membersList}
-            pairings={pairingsList}
-            setMembers={setMembersList}
-            setPairings={setPairingsList}
-          />
-        </Container>
-      </Grid>
-
-      {/* Table info for members/pairings */}
-      <Container minW="90%" centerContent mt={20}>
-        <Tabs w="100%">
-          <TabList>
-            <Tab>Members</Tab>
-            <Tab>Pairings</Tab>
-          </TabList>
-
-          <TabPanels>
-            {/* Members tab */}
-            <TabPanel>
-              <MembersTable
-                membersList={membersList}
-                changeMember={changeMember}
-                removeMember={removeMember}
-              />
-            </TabPanel>
-
-            {/* Pairings tab */}
-            <TabPanel>
-              <PairingsTable
-                pairings={pairingsList}
-                setPairings={setPairingsList}
-                members={membersList}
-                setMembers={setMembersList}
-                loading={loadingPairs}
-                refresh={refreshTables}
-              />
-            </TabPanel>
-          </TabPanels>
-        </Tabs>
-      </Container>
-    </Box>
-  );
-};
-
-export const getStaticProps: GetStaticProps = async () => {
-  // FIXME: use server side authentication to prevent unnecessary reads
-  // https://dev.to/theranbrig/server-side-authentication-with-nextjs-and-firebase-354m
-  const members: Member[] = await getAllMembers();
-  const pairings: Pairing[] = await getAllPairings();
-
-  return {
-    props: {
-      members,
-      pairings,
-    },
-    // TODO: learn more about incremental static regeneration .. but hopefully this helps
-    // https://nextjs.org/docs/basic-features/data-fetching#incremental-static-regeneration
-    revalidate: 5, // in seconds
-  };
-};
-
-export default EditPage;
